Products now on restock page
To be able to see anything on the
page aside from the navbar you need
to be logged in as an employee.

If logged in as an employee the
products are displayed on the page
with input fields that will be used
to edit price and stock (not implemented).

Issue 17